kong_admin_broker/requirements.txt
py4web>=1.20240906.1
psycopg2-binary>=2.9.9
requests>=2.31.0
pydal>=20231110.1
python-dotenv>=1.0.0
cryptography>=41.0.0
email-validator>=2.0.0